Remove reduction='sum' in SDAE loss function
This has the effect of scaling gradients by N * z, where N is the
batch size and z is the hidden layer. If we used SGD, we should
rescale the learning size to compensate for the much smaller losses
computed. If we used Adam without weight_decay or L2 regularization,
no changes to the hyperparamters would be needed. Since lambda_w
effectively acts as L2 regularization for SDAE, we use it as the
weight decay parameter for AdamW.